Fix daylight duration with flipped morning and evening times
If the evening time is before the morning time, QTime::secsTo() will
return a negative value. In that case, morning.secsTo(evening) is going
to actualy specify the time between evening and morning. For example,
witb the following times

  - evening: 1:00 AM
  - morning: 6:00 AM

morning.secsTo(evening) is going to return -18000, which is not what we
really expect. The issue can be addressed by applying a modulo operator.
But in order to make the code slightly more readable, this change goes
with an explicit "seconds in a day - time from evening to morning" formula.

BUG: 511293


(cherry picked from commit ce36931b53817c761540d099646b71c6458300d7)

Co-authored-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>